Moving ShopifyCli::Helpers::PartnersAPI to ShopifyCli::PartnersAPI - moving lib/shopify-cli/helpers/partners_api.rb to   lib/shopify-cli/partners_api.rb - moving test/shopify-cli/helpers/partners_api_test.rb to   test/shopify-cli/partners_api_test.rb - changing references of Helpers::PartnersAPI to ShopifyCli::PartnersAPI   in impacted helpers and tasks - changing autoload from lib/shopify-cli/helpers.rb to   lib/shopify_cli.rb
